requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.9.1
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.2